ADX-425 Explicitly add unaids specific schemas from a seperate module.
